pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.7",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
